Temporary fix for loss of pyserial support
The current version of FarmBot lost support for pyserial.  This is a temporary workaround that relies on a sleep call instead of waiting until it hears back from the second raspberry pi.  This code will be reverted to the original once the next version of FarmBot is released with pyserial.